### 2018-01-23 5c5ce5e
Update paper.md
