### 2018-01-23 #544 @arfon
👋 @Aneoshun - could you please merge this PR that fixes up the paper metadata?
 https://github.com/resibots/limbo/pull/249
