### 2018-01-23 #545 @whedon
[REVIEW]: Limbo: A Flexible High-performance Library for Gaussian Processes modeling and Data-Efficient Optimization
**Submitting author:** @aneoshun (<a href="http://orcid.org/0000-0002-3190-7073">Antoine Cully</a>)
**Repository:** <a href="https://github.com/resibots/limbo" target ="_blank">https://github.com/resibots/limbo</a>
**Version:** V2.0
