### 2018-03-05 37b2fa4
First version of GP with sparsification
