### 2018-03-05 044dff5
Added basic test
