### 2018-03-05 d9150a7
Added test for sparsification accuracy
