### 2018-03-07 93a79a7
Minor improvements
