### 2018-03-07 1b62a92
SparseGP --> SparsifiedGP
