### 2018-03-13 #545 @dfm
Hi team, thank you so much for you patience here! I've gone through this and I've included some comments below. This is a really impressive piece of software and I am excited to use it in my own work. I think that there are a few things that could improve the documentation make everything consistent with the JOSS guidelines, but it shouldn't be too onerous.

**Installation:**
- Overall, installation isn't too bad and the required dependencies can be easily installed following the directions here. Some of the optional deps were a bit annoying (classic C++ :smile:), but the docs here were complete.
- The `science` tap on Homebrew has been discontinued so `brew install homebrew/science/nlopt` throws an error. Perhaps the suggestion of using that should be removed.

**Performance:**
- There are strong claims in the paper when it comes to performance, but I didn't find that there was sufficient discussion on the benchmark pages of the docs to demonstrate this clearly. I think that it would be useful to include more discussion of what exactly is being tested in each experiment, why it matters, and a paragraph discussing the interpretation of the figures.
- It's not entirely obvious to me that GPy is the right benchmark for the GP comparisons. I'm not really sure what would be better, but there should (at least) be some discussion of the fact that GPy is *much* more feature rich and easy to use. The focus of GPy is not performance and it is written in pure-Python (+numpy).

**Documentation:**
- Overall the documentation is pretty complete, but I think that it would be much more useful if some of the earlier tutorial pages include more discussion of what is going on. Right now, the tutorials jump into source code pretty quickly and I think that a bit more theory (right on the tutorial pages!) would be useful.
- Specifically related to the JOSS requirements, some of the text from the paper could probably be added to the documentation home page to make the statement of need clearer.
- It would also be *very* useful to have examples of how to visualize the output of what is going on in each step of the optimization and tips on how to identify/debug models that aren't performing as expected.
- In the *Quick Start* example, `./waf --exp test` fails with `Cannot read the folder 'PATH_TO_LIMBO/limbo/exp/test'` for the correct value of `PATH_TO_LIMBO`.
- I think that the *Basic Example* needs a longer introduction to explain what it going on. In my experience, this is where users will try to start, and I think that adding more details here would go a long way.
- The full source code for the *Basic Example* includes the necessary `include`s and `namespace`, but the code snippets above should too.
- Again, I think that the *Basic Example* needs more discussion at the end about what to expect as output, how to interpret it, and how to visualize what it going on.
- The `Advanced Example` should probably have a listing of the full source code as well. Currently, the snippet for `eval_func` is missing the template and, even after fixing that, the code won't build on my machine (after I copied the listings directly). The error log is here: [advanced.log](https://github.com/openjournals/joss-reviews/files/1807983/advanced.log)

Thanks again for your patience! I hope that these comments are useful for improving the impact of this impressive library. Let me know if you have any questions.
