### 2018-03-14 92e4bce
Fixes for advanced example doc [ci skip]
