### 2018-03-14 cd8be66
Minor fix in basic example doc [ci skip]
