+++ src/limbo/spt/spt.hpp
@@ -1,0 +1,450 @@
+#ifndef LIMBO_SPT_SPT_HPP
+#define LIMBO_SPT_SPT_HPP
+
+#include <Eigen/Core>
+#include <Eigen/Dense>
+#include <algorithm>
+#include <limbo/tools/random_generator.hpp>
+#include <memory>
+#include <stack>
+#include <vector>
+
+namespace limbo {
+    namespace spt {
+
+        /// A implementation of Spatial Trees as described in
+        /// McFee, B. and Lanckriet, G.R., 2011. Large-scale music similarity search with spatial trees. In ISMIR (pp. 55-60).
+        struct SpatialTreeNode {
+        public:
+            SpatialTreeNode() : _parent(nullptr), _left(nullptr), _right(nullptr), _is_left(false), _depth(0) {}
+
+            SpatialTreeNode(const std::vector<Eigen::VectorXd>& points, const std::vector<Eigen::VectorXd>& observations)
+                : _parent(nullptr), _left(nullptr), _right(nullptr), _points(points), _observations(observations), _is_left(false), _depth(0) {}
+
+            const std::vector<Eigen::VectorXd>& points() const { return _points; }
+            const std::vector<Eigen::VectorXd>& observations() const { return _observations; }
+
+            double split_median() { return _split_median; }
+            void set_split_median(double sp_med) { _split_median = sp_med; }
+
+            double split_median_left() const { return _split_median_left; }
+            void set_split_median_left(double sp_med) { _split_median_left = sp_med; }
+
+            double split_median_right() const { return _split_median_right; }
+            void set_split_median_right(double sp_med) { _split_median_right = sp_med; }
+
+            const Eigen::VectorXd& split_dir() const { return _split_dir; }
+            void set_split_dir(const Eigen::VectorXd& sp_dir) { _split_dir = sp_dir; }
+
+            const Eigen::VectorXd& split_vector() const { return _split_vector; }
+            void set_split_vector(const Eigen::VectorXd& sp_vec) { _split_vector = sp_vec; }
+
+            const std::shared_ptr<SpatialTreeNode>& left() const { return _left; }
+            void set_left(const std::shared_ptr<SpatialTreeNode>& left) { _left = left; }
+
+            const std::shared_ptr<SpatialTreeNode>& right() const { return _right; }
+            void set_right(const std::shared_ptr<SpatialTreeNode>& right) { _right = right; }
+
+            const std::shared_ptr<SpatialTreeNode>& parent() const { return _parent; }
+            void set_parent(const std::shared_ptr<SpatialTreeNode>& parent) { _parent = parent; }
+
+            const Eigen::VectorXd& max() const { return _max; }
+            void set_max(const Eigen::VectorXd& max) { _max = max; }
+
+            const Eigen::VectorXd& min() const { return _min; }
+            void set_min(const Eigen::VectorXd& min) { _min = min; }
+
+            bool is_left() const { return _is_left; }
+            void make_left() { _is_left = true; }
+
+            size_t depth() const { return _depth; }
+            void set_depth(size_t d) { _depth = d; }
+
+            const std::vector<Eigen::VectorXd>& boundary_points() const { return _boundary_points; }
+            void set_boundary_points(const std::vector<Eigen::VectorXd>& boundary_points) { _boundary_points = boundary_points; }
+
+        protected:
+            std::shared_ptr<SpatialTreeNode> _parent, _left, _right;
+            std::vector<Eigen::VectorXd> _points, _observations, _boundary_points;
+            Eigen::VectorXd _split_dir, _split_vector;
+            Eigen::VectorXd _max, _min;
+            double _split_median, _split_median_left, _split_median_right;
+            bool _is_left;
[…]
