### 2018-03-20 b384ae0
A few more parameters for CMA-ES
