### 2018-03-27 65e9dc4
Added libGP to regression benchmarks
