+++ waf_tools/benchmark_template.cpp
@@ -7,6 +7,12 @@
 #include <limits>
 #include "test_functions.hpp"
 
+#ifdef USE_LIBGP
+#include <gp/gp.h>
+#include <gp/gp_utils.h>
+#include <gp/rprop.h>
+#endif
+
 using namespace limbo;
 
 template <typename T>
@@ -61,6 +67,9 @@ void benchmark(const std::string& name, std::vector<int> dimensions, std::vector
     int N_test = 10000;
 
     std::ofstream ofs_res(name + ".dat");
+#ifdef USE_LIBGP
+    std::ofstream ofs_libgp(name + "_libgp.dat");
+#endif
 
     for (size_t dim = 0; dim < dims.size(); dim++) {
         std::vector<Eigen::VectorXd> bounds = func.bounds();
@@ -76,6 +85,9 @@ void benchmark(const std::string& name, std::vector<int> dimensions, std::vector
 
             // Output number of models
             ofs_res << D << " " << N << " @NMODELS" << std::endl;
+#ifdef USE_LIBGP
+            ofs_libgp << D << " " << N << " 1" << std::endl;
+#endif
 
             std::cout << name << " in dim: " << D << " and # of points: " << N << std::endl;
             std::string file_name = name + "_" + std::to_string(D) + "_" + std::to_string(N);
@@ -107,6 +119,33 @@ void benchmark(const std::string& name, std::vector<int> dimensions, std::vector
                     obs[i] = obs[i].array() + gaussian_rand(0.0, sigma);
             }
 
+#ifdef USE_LIBGP
+            std::cout << "Training libGP GP..." << std::endl;
+            // Learn libGP GP here
+            libgp::GaussianProcess libgp_gp(D, "CovSum ( CovSEard, CovNoise)");
+            // set initial hyper-parameters
+            Eigen::VectorXd libgp_params = Eigen::VectorXd::Zero(libgp_gp.covf().get_param_dim());
+            libgp_params(D - 1) = std::log(0.01);
+            libgp_gp.covf().set_loghyper(libgp_params);
+            // add observations to libGP
+            for (int i = 0; i < N; i++) {
+                libgp_gp.add_pattern(points[i].data(), obs[i](0));
+            }
+            // optimize hyper-parameters of libGP
+            libgp::RProp rprop_optimizer;
+            double delta0 = 0.1;
+            double deltamin = 1e-6;
+            double deltamax = 50;
+            double etaminus = 0.5;
+            double etaplus = 1.2;
+            rprop_optimizer.init(1e-4, delta0, deltamin, deltamax, etaminus, etaplus);
+            auto start_libgp = std::chrono::high_resolution_clock::now();
+            rprop_optimizer.maximize(&libgp_gp, 300, false);
+            auto time_libgp = std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now() - start_libgp).count();
+            std::cout << std::setprecision(std::numeric_limits<long double>::digits10 + 1);
+            std::cout << "Time of libGP in secs: " << time_libgp / double(1000000.0) << std::endl;
+#endif
+
             // Learn the GPs code here
             @GPSLEARN
 
@@ -128,6 +167,32 @@ void benchmark(const std::string& name, std::vector<int> dimensions, std::vector
                 ob << func(p);
                 test_obs.push_back(ob);
             }
+
+#ifdef USE_LIBGP
+            // Predicition of the libGP GP
+            std::vector<double> predict_libgp(N_test);
+            std::vector<double> dummy_libgp(N_test);
+            start_libgp = std::chrono::high_resolution_clock::now();
+            for (int i = 0; i < N_test; i++) {
+                predict_libgp[i] = libgp_gp.f(test_points[i].data());
[…]
