+++ _sources/reg_benchmarks.rst.txt
@@ -1,7 +1,7 @@
 Gaussian process regression benchmarks
 ======================================
 
-*January 11, 2018* -- hal13 (32 cores)
+*March 28, 2018* -- hal14 (32 cores)
 
 - We compare to GPy (https://github.com/SheffieldML/GPy) 
 - Mean Squared Error: lower is better
