+++ _sources/bo_benchmarks.rst.txt
@@ -1,7 +1,7 @@
 Bayesian optimization benchmarks
 ===============================
 
-*January 10, 2018* -- hal23 (32 cores)
+*March 28, 2018* -- hal13 (32 cores)
 
 - We compare to BayesOpt (https://github.com/rmcantin/bayesopt) 
 - Accuracy: lower is better (difference with the optimum)
