### 2018-04-05 5f68546
fix the flier & box colors in plot_bo_benchmarks
