### 2018-04-06 a586f13
clean labels in plot_bo_benchmarks
