+++ v2.0.0/_sources/reg_benchmarks.rst.txt
@@ -1,7 +1,7 @@
 Gaussian process regression benchmarks
 ======================================
 
-*March 30, 2018* -- hal13 (32 cores)
+*April 04, 2018* -- hal14 (32 cores)
 
 - We compare to GPy (https://github.com/SheffieldML/GPy) and libGP (https://github.com/mblum/libgp)
 - Mean Squared Error: lower is better
