### 2018-04-16 4f1191d
add notes for benchmark pages in docs.
