+++ v1.0.0/_sources/bo_benchmarks.rst.txt
@@ -17,7 +17,7 @@ Details
 - see `src/benchmarks/limbo/bench.cpp` and `src/benchmarks/bayesopt/bench.cpp`
 
 
-branin
+Branin function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/branin.html>`_. 
@@ -26,7 +26,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/bran
 
 .. figure:: fig_benchmarks/branin.png
 
-ellipsoid
+Ellipsoid function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/rothyp.html>`_. 
@@ -35,7 +35,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/roth
 
 .. figure:: fig_benchmarks/ellipsoid.png
 
-goldsteinprice
+Goldsteinprice function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/goldpr.html>`_. 
@@ -44,7 +44,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/gold
 
 .. figure:: fig_benchmarks/goldsteinprice.png
 
-hartmann3
+Hartmann3 function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/hart3.html>`_. 
@@ -53,7 +53,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/hart
 
 .. figure:: fig_benchmarks/hartmann3.png
 
-hartmann6
+Hartmann6 function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/hart6.html>`_. 
@@ -62,7 +62,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/hart
 
 .. figure:: fig_benchmarks/hartmann6.png
 
-rastrigin
+Rastrigin function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/rastr.html>`_ (note, the input space has been rescaled and shifted to have the minimum at [1,1,1,1], instead of [0,0,0,0] 
@@ -71,7 +71,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/rast
 
 .. figure:: fig_benchmarks/rastrigin.png
 
-sixhumpcamel
+Sixhumpcamel function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/camel6.html>`_. 
@@ -80,7 +80,7 @@ Details about the function can be found `here <https://www.sfu.ca/~ssurjano/came
 
 .. figure:: fig_benchmarks/sixhumpcamel.png
 
-sphere
+Sphere function
 -----------------
 
 Details about the function can be found `here <https://www.sfu.ca/~ssurjano/spheref.html>`_. 
