### 2018-05-10 e39c384
Fixes in benchmarks [ci skip]
