+++ waf_tools/plot_regression_benchmarks.py
@@ -81,7 +81,7 @@ params = {
 rcParams.update(params)
 
 def load_data():
-    files = glob("regression_benchmark_results/*/*/*.dat")
+    files = sorted(glob("regression_benchmark_results/*/*/*.dat"))
     points = defaultdict(lambda : defaultdict(lambda : defaultdict(dict)))
     times_learn = defaultdict(lambda : defaultdict(lambda : defaultdict(dict)))
     times_query = defaultdict(lambda : defaultdict(lambda : defaultdict(dict)))
@@ -159,7 +159,7 @@ def plot_ax(ax, data, points, labely, disp_legend=True, disp_xaxis=False):
             if var_p[i] not in pp:
                 pp[var_p[i]] = []
             pp[var_p[i]].append(var_data[i])
-        
+
         pp = OrderedDict(sorted(pp.items()))
 
         x_axis = pp.keys()
@@ -178,7 +178,7 @@ def plot_ax(ax, data, points, labely, disp_legend=True, disp_xaxis=False):
         ax.plot(x_axis, y_axis, '-o', color=c_kk, linewidth=3, markersize=5)
         ax.fill_between(x_axis, y_axis_75, y_axis_25, color=c_kk, alpha=0.15, linewidth=2)
         kk = kk + 1
-    
+
     if disp_legend:
         ax.legend(labels)
     if disp_xaxis:
@@ -189,19 +189,27 @@ def plot_ax(ax, data, points, labely, disp_legend=True, disp_xaxis=False):
     return replicates
 
 def get_notes():
-    notes={'Rastrigin':'Details about the function can be found `here <https://www.sfu.ca/~ssurjano/rastr.html>`_.',
-           'Pistonsimulation':'Details about the function can be found `here <https://www.sfu.ca/~ssurjano/piston.html>`_.',
-           'Step':'Step function: for :math:`x\in[-2; 2]` :\n.. math::\n  f(x) = \\begin{cases} 0, &\\mbox{if x<=0} \\\\ 1, &\\mbox{otherwise} \\end{cases}',
-           'Robotarm':'Details about the function can be found `here <https://www.sfu.ca/~ssurjano/robot.html>`_.',
-           'Gramacylee':'Details about the function can be found `here <https://www.sfu.ca/~ssurjano/grlee12.html>`_.',
-           'Planarinversedynamicsii':'Approximation of the second motor\'s torque in the inverse dynamics of a Planar 2D Arm. Details are given at the bottom of this page.',
-           'Planarinversedynamicsi':'Approximation of the first motor\'s torque in the inverse dynamics of a Planar 2D Arm. Details are given at the bottom of this page.',
-           'Otlcircuit':'Details about the function can be found `here <https://www.sfu.ca/~ssurjano/otlcircuit.html>`_.'};
+    notes={'Rastrigin':"Details about the function can be found `here <https://www.sfu.ca/~ssurjano/rastr.html>`_.",
+           'Pistonsimulation':"Details about the function can be found `here <https://www.sfu.ca/~ssurjano/piston.html>`_.",
+           'Step':"Step function: for :math:`x\in[-2; 2]` :\n\n.. math::\n  f(x) = \\begin{cases} 0, &\\mbox{if x<=0} \\\\ 1, &\\mbox{otherwise} \\end{cases}",
+           'Robotarm':"Details about the function can be found `here <https://www.sfu.ca/~ssurjano/robot.html>`_.",
+           'Gramacylee':"Details about the function can be found `here <https://www.sfu.ca/~ssurjano/grlee12.html>`_.",
+           'Planarinversedynamicsii':"Approximation of the second motor\'s torque in the inverse dynamics of a Planar 2D Arm. Details are given at the bottom of this page.",
[…]
