### 2018-05-10 8304897
add incremental updates for SSGPR
