+++ src/limbo/experimental/model/ssgpr.hpp
@@ -2,6 +2,7 @@
 #define LIMBO_EXPERIMENTAL_MODEL_SSPGR_HPP
 
 #include <vector>
+#include <chrono>
 
 #include <Eigen/Cholesky>
 #include <Eigen/Core>
@@ -32,10 +33,10 @@ namespace limbo {
                 {
                     assert(_dim_in > 0 && _dim_out > 0);
 
-                    int n = _mapping.dim_out();
-                    _L = _sigma_n * _sigma_n * Eigen::MatrixXd::Identity(n, n);
-                    _B = Eigen::MatrixXd::Zero(n, _dim_out);
-                    _W = Eigen::MatrixXd::Zero(n, _dim_out);
+                    _dim_mapping = _mapping.dim_out();
+                    _L = _sigma_n * Eigen::MatrixXd::Identity(_dim_mapping, _dim_mapping);
+                    _B = Eigen::MatrixXd::Zero(_dim_mapping, _dim_out);
+                    _W = Eigen::MatrixXd::Zero(_dim_mapping, _dim_out);
                 }
 
                 std::tuple<Eigen::VectorXd, double> query(const Eigen::VectorXd& x) const
@@ -92,10 +93,10 @@ namespace limbo {
 
                     this->_compute_obs_mean();
                     if (compute_full)
-                        this->_full_compute();
+                        this->_compute_full_kernel();
                 }
 
-                /// Do not forget to call this if you use hyper-prameters optimization!!
+                /// Do not forget to call this if you use hyper-parameters optimization!!
                 void optimize_hyperparams()
                 {
                     _hp_optimize(*this);
@@ -109,7 +110,7 @@ namespace limbo {
                     if (update_obs_mean)
                         this->_compute_obs_mean();
 
-                    this->_full_compute();
+                    this->_compute_full_kernel();
                 }
 
                 /// return the number of samples used to compute the GP
@@ -169,8 +170,59 @@ namespace limbo {
                     _sigma_n = std::exp(p(0)); //p(0); //std::exp(p(0)) + 1e-6;
                 }
 
+            /// add sample and update the GP. This code uses an incremental implementation of the Cholesky
+            /// decomposition. It is therefore much faster than a call to compute()
+            void add_sample(const Eigen::VectorXd& sample, const Eigen::VectorXd& observation)
+            {
+                if (_samples.empty()) {
+                    if (_dim_in != sample.size()) {
+                        _dim_in = sample.size();
+                        _mapping = Mapping(_dim_in);
+                        reset();
+                    }
+                    if (_dim_out != observation.size()) {
+                        _dim_out = observation.size();
+                        _mean_function = MeanFunction(_dim_out); // the cost of building a functor should be relatively low
+                        reset();
+                    }
+                }
+                else {
+                    assert(sample.size() == _dim_in);
+                    assert(observation.size() == _dim_out);
+                }
+
+                _samples.push_back(sample);
+
+                _observations.conservativeResize(_observations.rows()+1, _dim_out);
+                _observations.bottomRows<1>() = observation.transpose();
+                // _mean_observation = _observations.colwise().mean();
+
+                this->_compute_obs_mean();
+
+                _Phi.conservativeResize(_Phi.rows()+1, _mapping.dim_out());
+                Eigen::VectorXd sample_phi = _mapping.evaluate(sample);
+                _Phi.bottomRows<1>() = sample_phi;
+
+                // this->_compute_incremental_kernel(); 
+
+                // Make Rtilde - [R, sample_phi]
+                Eigen::MatrixXd Rtilde(_dim_mapping+1, _dim_mapping);
+                Rtilde.topRows(_dim_mapping) = _L.transpose();
[…]
