### 2018-05-10 8528b85
typo fixes and clarification
