+++ src/limbo/model/sparsified_gp.hpp
@@ -76,7 +76,7 @@ namespace limbo {
             /// useful because the model might be created before knowing anything about the process
             SparsifiedGP() : base_gp_t() {}
 
-            /// useful because the model might be created  before having samples
+            /// useful because the model might be created before having samples
             SparsifiedGP(int dim_in, int dim_out)
                 : base_gp_t(dim_in, dim_out) {}
 
