### 2018-05-11 8d075b6
Avoid making extra matrix in incremental update
