### 2018-05-14 9ffa659
[ci skip] minor improvement basic example
