+++ fix_docs/_sources/defaults.rst.txt
@@ -1,0 +1,156 @@
+Default values
+===============
+kernel
+^^^^^^
+- bool optimize_noise = false [defined in src/limbo/kernel/kernel.hpp]
+- double noise = 0.01 [defined in src/limbo/kernel/kernel.hpp]
+
+kernel_maternthreehalves
+^^^^^^^^^^^^^^^^^^^^^^^^
+- double sigma_sq = 1 [defined in src/limbo/kernel/matern_three_halves.hpp]
+- double l = 1 [defined in src/limbo/kernel/matern_three_halves.hpp]
+
+acqui_eci
+^^^^^^^^^
+- double jitter = 0.0 [defined in src/limbo/experimental/acqui/eci.hpp]
+
+bayes_opt_bobase
+^^^^^^^^^^^^^^^^
+- bool stats_enabled = true [defined in src/limbo/bayes_opt/bo_base.hpp]
+- bool bounded = true [defined in src/limbo/bayes_opt/bo_base.hpp]
+
+acqui_ucb_imgpo
+^^^^^^^^^^^^^^^
+- double nu = 0.05 [defined in src/limbo/experimental/acqui/ucb_imgpo.hpp]
+
+model_sparse_gp
+^^^^^^^^^^^^^^^
+- int max_points = 200 [defined in src/limbo/model/sparsified_gp.hpp]
+
+init_lhs
+^^^^^^^^
+- int samples = 10 [defined in src/limbo/init/lhs.hpp]
+
+init_randomsamplinggrid
+^^^^^^^^^^^^^^^^^^^^^^^
+- int bins = 5 [defined in src/limbo/init/random_sampling_grid.hpp]
+- int samples = 10 [defined in src/limbo/init/random_sampling_grid.hpp]
+
+mean_constant
+^^^^^^^^^^^^^
+- double constant = 1 [defined in src/limbo/mean/constant.hpp]
+
+model_spgp
+^^^^^^^^^^
+- double jitter = 0.000001 [defined in src/limbo/experimental/model/spgp.hpp]
+- double samples_percent = 10 [defined in src/limbo/experimental/model/spgp.hpp]
+- int min_m = 1 [defined in src/limbo/experimental/model/spgp.hpp]
+- double pred_kernel_sigma_sq = 0.5 [defined in src/limbo/experimental/model/spgp.hpp]
+- double sig = 0.01 [defined in src/limbo/experimental/model/spgp.hpp]
+- double pred_kernel_l = 0.5 [defined in src/limbo/experimental/model/spgp.hpp]
+
+stop_maxpredictedvalue
+^^^^^^^^^^^^^^^^^^^^^^
+- double ratio = 0.9 [defined in src/limbo/stop/max_predicted_value.hpp]
+
+stop_maxiterations
+^^^^^^^^^^^^^^^^^^
+- int iterations = 190 [defined in src/limbo/stop/max_iterations.hpp]
+
+opt_rprop
+^^^^^^^^^
+- double eps_stop = 0.0 [defined in src/limbo/opt/rprop.hpp]
+- int iterations = 300 [defined in src/limbo/opt/rprop.hpp]
+
+bayes_opt_boptimizer
+^^^^^^^^^^^^^^^^^^^^
+- int hp_period = -1 [defined in src/limbo/bayes_opt/boptimizer.hpp]
+
+kernel_maternfivehalves
+^^^^^^^^^^^^^^^^^^^^^^^
+- double sigma_sq = 1 [defined in src/limbo/kernel/matern_five_halves.hpp]
[…]
