+++ release-2.0/_sources/reg_benchmarks.rst.txt
@@ -1,7 +1,7 @@
 Gaussian process regression benchmarks
 ======================================
 
-*May 11, 2018* -- hal27 (32 cores)
+*May 16, 2018* -- hal14 (32 cores)
 
 This page presents benchmarks in which we compare the performance of the Gaussian Process regression in **Limbo** against two other libraries: GPy (https://github.com/SheffieldML/GPy) and libGP (https://github.com/mblum/libgp).
 
