### 2018-05-25 e790ea5
[ci skip] Add minimal matplotlib script to display aggredated observations
