### 2018-05-31 9d50b75
Minor typo in rprop docs [ci skip]
