### 2018-06-01 ce836b6
automatic update of the doc [skip ci]
