+++ v2.0.0/_sources/bo_benchmarks.rst.txt
@@ -1,7 +1,7 @@
 Bayesian optimization benchmarks
 ===============================
 
-*May 16, 2018* -- hal13 (32 cores)
+*May 30, 2018* -- hal13 (32 cores)
 
 This page presents benchmarks in which we compare the Bayesian optimization performance of **Limbo** against BayesOpt (https://github.com/rmcantin/bayesopt , a state-of-the-art Bayesian Optimization library in C++).
 
