### 2018-06-01 #545 @Aneoshun
Dear @dfm and @arfon,

Thank you very much for your comments and your patience. We have addressed all of them and we believe that we made the documentation better.

You can see the changes that we made in this pull request: https://github.com/resibots/limbo/pull/257.  Please, also find bellow our response to your comments:

Installation:
* Overall, installation isn't too bad and the required dependencies can be easily installed following the directions here. Some of the optional deps were a bit annoying (classic C++ 😄), but the docs here were complete.
    * We are happy to read this.

* The science tap on Homebrew has been discontinued so brew install homebrew/science/nlopt throws an error. Perhaps the suggestion of using that should be removed.
    * We have changed the documentation and advised to install nlopt via: brew install nlopt (http://www.resibots.eu/limbo/tutorials/compilation.html)

Performance:
* There are strong claims in the paper when it comes to performance, but I didn't find that there was sufficient discussion on the benchmark pages of the docs to demonstrate this clearly. I think that it would be useful to include more discussion of what exactly is being tested in each experiment, why it matters, and a paragraph discussing the interpretation of the figures.
    * We have significantly extended and restructured the discussions on the two benchmark pages (http://www.resibots.eu/limbo/reg_benchmarks.html  and http://www.resibots.eu/limbo/bo_benchmarks.html). In particular, we have added several paragraphs to introduce the benchmarks and give more details about the compared methods.

* It's not entirely obvious to me that GPy is the right benchmark for the GP comparisons. I'm not really sure what would be better, but there should (at least) be some discussion of the fact that GPy is much more feature rich and easy to use. The focus of GPy is not performance and it is written in pure-Python (+numpy).
    * We have added another library in our regression benchmark: LibGP, which is a C++ library for Gaussian Processes (https://github.com/mblum/libgp). We also included in the discussion of this page (http://www.resibots.eu/limbo/reg_benchmarks.html), a paragraph that explains that GPy is a python library with much more feature and designed to be easy to use. Moreover, GPy can achieve comparable performance with C++ libraries in the hyper-parameters optimization part because it utilizes numpy and scipy that is basically calling C code with MKL bindings (which is almost identical to what we are doing in Limbo).

Documentation:
* Overall the documentation is pretty complete, but I think that it would be much more useful if some of the earlier tutorial pages include more discussion of what is going on. Right now, the tutorials jump into source code pretty quickly and I think that a bit more theory (right on the tutorial pages!) would be useful.
    *  We offer in the document a quick introduction to Bayesian Optimization in the Guide “ Introduction to Bayesian Optimization (BO)” (http://www.resibots.eu/limbo/guides/bo.html) in which we introduce the main concepts of BO. However, we agree that such concepts need to be known before starting the tutorial of Limbo. For this reason, and following your suggestions, we added a sentence at the beginning of the basic_example that invite interested readers to refer to this introduction (http://www.resibots.eu/limbo/tutorials/basic_example.html).

* Specifically related to the JOSS requirements, some of the text from the paper could probably be added to the documentation home page to make the statement of need clearer.
    * We have added several paragraphs from the paper in the home page of the documentation following your advice (http://www.resibots.eu/limbo/index.html).

* It would also be very useful to have examples of how to visualize the output of what is going on in each step of the optimization and tips on how to identify/debug models that aren't performing as expected.
    * We have included in the basic example a small python script to display the values observed at iterations by Limbo (http://www.resibots.eu/limbo/tutorials/basic_example.html).

* In the Quick Start example, ./waf --exp test fails with Cannot read the folder 'PATH_TO_LIMBO/limbo/exp/test' for the correct value of PATH_TO_LIMBO.
    * We have not been able to reproduce this problem. Are you sure that you ran the command: `./waf --create test` before attempting to compile (as indicated in the documentation)?

* I think that the Basic Example needs a longer introduction to explain what it going on. In my experience, this is where users will try to start, and I think that adding more details here would go a long way.
    * We have added several lines to introduce better the Basic Example. In particular, we highlight the objective of this example and links to the definition of the main concepts (http://www.resibots.eu/limbo/tutorials/basic_example.html).

* The full source code for the Basic Example includes the necessary includes and namespace, but the code snippets above should too.
    * Thank you for pointing this out. We have modified the Basic Example page accordingly (http://www.resibots.eu/limbo/tutorials/basic_example.html).

* Again, I think that the Basic Example needs more discussion at the end about what to expect as output, how to interpret it, and how to visualize what it going on.
    * We have extended the Basic Example to include a typical result obtained after running the example. In particular, we details the files that are produced by Limbo and how to interpret them (http://www.resibots.eu/limbo/tutorials/basic_example.html).

* The Advanced Example should probably have a listing of the full-time source code as well. Currently, the snippet for eval_func is missing the template and, even after fixing that, the code won't build on my machine (after I copied the listings directly). The error log is here: advanced.log
    * We have fixed the issues you mentioned and provided the full code of the Advance Example (http://www.resibots.eu/limbo/tutorials/advanced_example.html).

In general, we believe that your comments really helped us to improve the documentation and the library. We hope you will like these changes.

Best regards,
