### 2018-06-10 5063765
Adding a few gradient-based optimizers
