### 2018-06-19 b9ef3bb
Add possibility of progress functions in CMA-ES
