### 2018-06-21 ee62dd3
Fix sigma^2 prediction of GPs
