### 2018-06-21 5c76870
Fix gp tests for new sigma
