### 2018-06-25 1d4cd40
No need to have parallel repeater as default HP optimizer
