### 2018-06-25 6c82699
Added MultiGP - solves #266
