+++ src/tests/test_gp.cpp
@@ -61,6 +61,8 @@
 #include <limbo/model/gp/kernel_loo_opt.hpp>
 #include <limbo/model/gp/kernel_mean_lf_opt.hpp>
 #include <limbo/model/gp/mean_lf_opt.hpp>
+#include <limbo/model/multi_gp.hpp>
+#include <limbo/model/multi_gp/parallel_lf_opt.hpp>
 #include <limbo/model/sparsified_gp.hpp>
 #include <limbo/opt/grid_search.hpp>
 #include <limbo/tools/macros.hpp>
@@ -905,4 +907,208 @@ BOOST_AUTO_TEST_CASE(test_sparse_gp_accuracy)
     }
 
     BOOST_CHECK(double(failures) / double(N) < 0.1);
+}
+
+BOOST_AUTO_TEST_CASE(test_multi_gp_dim)
+{
+    using namespace limbo;
+
+    using KF_t = kernel::MaternFiveHalves<Params>;
+    using Mean_t = mean::Constant<Params>;
+    using GP_t = model::GP<Params, KF_t, Mean_t>;
+
+    GP_t gp; // no init with dim
+
+    std::vector<Eigen::VectorXd> observations = {make_v2(5, 5), make_v2(10, 10),
+        make_v2(5, 5)};
+    std::vector<Eigen::VectorXd> samples = {make_v2(1, 1), make_v2(2, 2), make_v2(3, 3)};
+
+    gp.compute(samples, observations);
+
+    using MultiGP_t = model::MultiGP<Params, model::GP, KF_t, Mean_t>;
+    MultiGP_t multi_gp;
+    multi_gp.compute(samples, observations);
+
+    Eigen::VectorXd mu;
+    double sigma;
+    std::tie(mu, sigma) = gp.query(make_v2(1, 1));
+    BOOST_CHECK(std::abs((mu(0) - 5)) < 1);
+    BOOST_CHECK(std::abs((mu(1) - 5)) < 1);
+
+    BOOST_CHECK(sigma <= 2. * (Params::kernel::noise() + 1e-8));
+
+    Eigen::VectorXd mu_multi, sigma_multi;
+
+    std::tie(mu_multi, sigma_multi) = multi_gp.query(make_v2(1, 1));
+    BOOST_CHECK(std::abs((mu_multi(0) - 5)) < 1);
+    BOOST_CHECK(std::abs((mu_multi(1) - 5)) < 1);
+
+    BOOST_CHECK(sigma_multi(0) <= 2. * (Params::kernel::noise() + 1e-8));
+    BOOST_CHECK(sigma_multi(1) <= 2. * (Params::kernel::noise() + 1e-8));
+
+    BOOST_CHECK(std::abs(mu_multi(0) - mu(0)) < 1e-6);
+    BOOST_CHECK(std::abs(mu_multi(1) - mu(1)) < 1e-6);
+    BOOST_CHECK(std::abs(sigma_multi(0) - sigma) < 1e-6);
+    BOOST_CHECK(std::abs(sigma_multi(1) - sigma) < 1e-6);
+}
+
+BOOST_AUTO_TEST_CASE(test_multi_gp)
+{
+    using namespace limbo;
+
+    using KF_t = kernel::MaternFiveHalves<Params>;
+    using Mean_t = mean::Constant<Params>;
+    using MultiGP_t = model::MultiGP<Params, model::GP, KF_t, Mean_t>;
+
+    MultiGP_t gp;
+    std::vector<Eigen::VectorXd> observations = {make_v1(5), make_v1(10),
+        make_v1(5)};
+    std::vector<Eigen::VectorXd> samples = {make_v1(1), make_v1(2), make_v1(3)};
[…]
