### 2018-06-25 1850c5c
Added test for multi-gp recompute
